requirements.txt
fastapi==0.115.13
django==4.2.10
djangorestframework==3.14.0
psycopg[binary]==3.2.9
python-dotenv==1.0.0
dj-database-url==1.0.0
django-cors-headers==4.0.0